Fix DebugKit integration
When DebugKit's `SqlLogPanel` kicks in, it assumes that every connection
is of an SQL type and there is no way to just tell it to ignore that.

For now, I have decided to try leveraging that but in case it doesn't
work out good, might be good to think about keeping registering a custom
log panel as part of the plugin.